Move stripUsername to separate module + other changes
